Banco com 1000 registros de hospitalização do Equador.

A Encuesta de Egresos Hospitalarios registra as hospitalizações ocorridas em todo o Equador, no setor público e privado. Aqui é apresentada uma amostra aleatória da base de dados do ano 2020.

Usage

data(eeh20)

Format

An object of class tbl_df (inherits from tbl, data.frame) with 1000 rows and 36 columns.

Source

https://www.ecuadorencifras.gob.ec/

Examples

data("eeh20")
str(eeh20)
#> tibble [1,000 × 36] (S3: tbl_df/tbl/data.frame)
#>  $ prov_ubi  : chr+lbl [1:1000] 09, 09, 09, 12, 09, 09, 09, 17, 09, 12, 17, 11, 18, 1...
#>    ..@ label        : chr "Provincia ubicación establecimiento"
#>    ..@ format.spss  : chr "A6"
#>    ..@ display_width: int 2
#>    ..@ labels       : Named chr [1:25] "10" "20" "01" "11" ...
#>    .. ..- attr(*, "names")= chr [1:25] "Imbabura" "Galápagos" "Azuay" "Loja" ...
#>  $ cant_ubi  : chr+lbl [1:1000] 0901, 0901, 0901, 1207, 0901, 0901, 0901, 1701, 0901,...
#>    ..@ label        : chr "SecA Cantón ubicación establecimiento"
#>    ..@ format.spss  : chr "A12"
#>    ..@ display_width: int 14
#>    ..@ labels       : Named chr [1:222] "8800" "0110" "1110" "1210" ...
#>    .. ..- attr(*, "names")= chr [1:222] "Exterior" "Oña" "Puyango" "Buena Fé" ...
#>  $ parr_ubi  : chr+lbl [1:1000] 090114, 090114, 090112, 120750, 090112, 090112, 09011...
#>    ..@ label        : chr "SecA Parroquia ubicación establecimiento"
#>    ..@ format.spss  : chr "A18"
#>    ..@ display_width: int 20
#>    ..@ labels       : Named chr [1:1308] "880000" "010110" "170110" "090110" ...
#>    .. ..- attr(*, "names")= chr [1:1308] "Exterior" "San Blas" "El Condado" "Rocafuerte" ...
#>  $ area_ubi  : dbl+lbl [1:1000] 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,...
#>    ..@ label      : chr "Area de ubicación del establecimiento"
#>    ..@ format.spss: chr "F1.0"
#>    ..@ labels     : Named num [1:2] 1 2
#>    .. ..- attr(*, "names")= chr [1:2] "Urbana" "Rural"
#>  $ clase     : dbl+lbl [1:1000] 11,  2,  5,  1,  1,  2,  1,  2,  5,  1, 11,  2,  2,  ...
#>    ..@ label      : chr "Clase de establecimiento"
#>    ..@ format.spss: chr "F2.0"
#>    ..@ labels     : Named num [1:18] 1 2 3 4 5 6 7 8 9 10 ...
#>    .. ..- attr(*, "names")= chr [1:18] "Hospital básico" "Hospital general" "Infectología" "Gineco-Obstétrico" ...
#>  $ tipo      : dbl+lbl [1:1000] 1, 1, 1, 5, 5, 1, 5, 1, 1, 5, 1, 1, 1, 5, 1, 1, 1, 1,...
#>    ..@ label      : chr "Tipo de establecimiento"
#>    ..@ format.spss: chr "F1.0"
#>    ..@ labels     : Named num [1:5] 1 2 3 4 5
#>    .. ..- attr(*, "names")= chr [1:5] "Agudo" "Crónico" "Clínicas generales sin especialidad" "Establecimientos sin internación" ...
#>  $ entidad   : dbl+lbl [1:1000]  6,  1, 13,  1, 18,  1, 18,  1,  1,  1,  6,  1,  6, 1...
#>    ..@ label      : chr "Entidad a la que pertenece el establecimiento"
#>    ..@ format.spss: chr "F2.0"
#>    ..@ labels     : Named num [1:18] 1 2 3 4 5 6 7 8 9 10 ...
#>    .. ..- attr(*, "names")= chr [1:18] "Ministerio de Salud Pública" "Ministerio de Justicia, Derechos Humanos y Cultos" "Ministerio de Defensa Nacional" "Ministerio de Educación" ...
#>  $ sector    : dbl+lbl [1:1000] 1, 1, 3, 1, 2, 1, 2, 1, 1, 1, 1, 1, 1, 2, 1, 2, 1, 1,...
#>    ..@ label      : chr "Sector al que pertenece el establecimiento"
#>    ..@ format.spss: chr "F3.0"
#>    ..@ labels     : Named num [1:3] 1 2 3
#>    .. ..- attr(*, "names")= chr [1:3] "Público" "Privado con fines de lucro" "Privado sin fines de lucro"
#>  $ mes_inv   : dbl+lbl [1:1000]  9, 11,  2,  8,  8,  3, 12,  7, 12,  2,  1, 11,  6,  ...
#>    ..@ label      : chr "Mes de registro/investigación"
#>    ..@ format.spss: chr "F2.0"
#>    ..@ labels     : Named num [1:12] 1 2 3 4 5 6 7 8 9 10 ...
#>    .. ..- attr(*, "names")= chr [1:12] "Enero" "Febrero" "Marzo" "Abril" ...
#>  $ nac_pac   : dbl+lbl [1:1000] 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 1,...
#>    ..@ label      : chr "Nacionalidad del paciente"
#>    ..@ format.spss: chr "F1.0"
#>    ..@ labels     : Named num [1:3] 1 2 9
#>    .. ..- attr(*, "names")= chr [1:3] "Ecuatoriano/a" "Extranjero/a" "Ignorado"
#>  $ cod_pais  : num [1:1000] 218 218 218 218 218 218 218 218 218 218 ...
#>   ..- attr(*, "label")= chr "Código del país"
#>   ..- attr(*, "format.spss")= chr "F3.0"
#>  $ nom_pais  : chr [1:1000] "ECUADOR" "ECUADOR" "ECUADOR" "ECUADOR" ...
#>   ..- attr(*, "label")= chr "Nombre de país"
#>   ..- attr(*, "format.spss")= chr "A21"
#>   ..- attr(*, "display_width")= int 7
#>  $ sexo      : dbl+lbl [1:1000] 1, 2, 1, 2, 2, 2, 1, 2, 2, 1, 1, 2, 1, 2, 1, 1, 2, 2,...
#>    ..@ label      : chr "Sexo del paciente"
#>    ..@ format.spss: chr "F1.0"
#>    ..@ labels     : Named num [1:3] 1 2 3
#>    .. ..- attr(*, "names")= chr [1:3] "Hombre" "Mujer" "Indeterminado"
#>  $ cod_edad  : dbl+lbl [1:1000] 4, 4, 4, 4, 4, 4, 4, 4, 2, 4, 4, 4, 4, 4, 4, 4, 4, 4,...
#>    ..@ label      : chr "Condición de la edad del paciente"
#>    ..@ format.spss: chr "F1.0"
#>    ..@ labels     : Named num [1:5] 1 2 3 4 9
#>    .. ..- attr(*, "names")= chr [1:5] "Horas (1 a 23 horas de edad)" "Días (1 a 28 días de edad)" "Meses (1 a 11 meses de edad)" "Años (1 a 115 años de edad)" ...
#>  $ edad      : num [1:1000] 52 57 9 24 56 17 36 20 10 67 ...
#>   ..- attr(*, "label")= chr "Edad del paciente"
#>   ..- attr(*, "format.spss")= chr "F3.0"
#>  $ etnia     : dbl+lbl [1:1000] 8, 6, 6, 6, 6, 6, 6, 6, 6, 6, 9, 6, 9, 6, 6, 6, 6, 6,...
#>    ..@ label      : chr "Definicion étnica del paciente"
#>    ..@ format.spss: chr "F1.0"
#>    ..@ labels     : Named num [1:9] 1 2 3 4 5 6 7 8 9
#>    .. ..- attr(*, "names")= chr [1:9] "Indígena" "Afroecuatoriano/a Afrodescendiente" "Negro/a" "Mulato/a" ...
#>  $ prov_res  : chr+lbl [1:1000] 09, 09, 09, 12, 09, 09, 09, 17, 09, 12, 17, 11, 18, 1...
#>    ..@ label        : chr "Provincia de residencia habitual del paciente"
#>    ..@ format.spss  : chr "A6"
#>    ..@ display_width: int 2
#>    ..@ labels       : Named chr [1:25] "10" "20" "01" "11" ...
#>    .. ..- attr(*, "names")= chr [1:25] "Imbabura" "Galápagos" "Azuay" "Loja" ...
#>  $ area_res  : num [1:1000] 1 1 1 1 1 1 1 2 1 1 ...
#>   ..- attr(*, "label")= chr "Area de residencia habitual del paciente"
#>   ..- attr(*, "format.spss")= chr "F1.0"
#>  $ anio_ingr : num [1:1000] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 ...
#>   ..- attr(*, "label")= chr "Año de ingreso"
#>   ..- attr(*, "format.spss")= chr "F4.0"
#>  $ mes_ingr  : dbl+lbl [1:1000]  8, 10,  2,  8,  8,  3, 12,  7, 12,  2,  1, 11,  6,  ...
#>    ..@ label      : chr "Mes de ingreso"
#>    ..@ format.spss: chr "F2.0"
#>    ..@ labels     : Named num [1:12] 1 2 3 4 5 6 7 8 9 10 ...
#>    .. ..- attr(*, "names")= chr [1:12] "Enero" "Febrero" "Marzo" "Abril" ...
#>  $ dia_ingr  : num [1:1000] 30 30 14 12 6 4 16 17 15 18 ...
#>   ..- attr(*, "label")= chr "Día de ingreso"
#>   ..- attr(*, "format.spss")= chr "F2.0"
#>  $ fecha_ingr: chr [1:1000] "2020/08/30 00:00:00.000" "2020/10/30 00:00:00.000" "2020/02/14 00:00:00.000" "2020/08/12 00:00:00.000" ...
#>   ..- attr(*, "label")= chr "Fecha de egreso"
#>   ..- attr(*, "format.spss")= chr "A69"
#>   ..- attr(*, "display_width")= int 23
#>  $ anio_egr  : num [1:1000] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 ...
#>   ..- attr(*, "label")= chr "Año del egreso"
#>   ..- attr(*, "format.spss")= chr "F4.0"
#>  $ mes_egr   : dbl+lbl [1:1000]  9, 11,  2,  8,  8,  3, 12,  7, 12,  2,  1, 11,  6,  ...
#>    ..@ label      : chr " Mes egreso"
#>    ..@ format.spss: chr "F2.0"
#>    ..@ labels     : Named num [1:12] 1 2 3 4 5 6 7 8 9 10 ...
#>    .. ..- attr(*, "names")= chr [1:12] "Enero" "Febrero" "Marzo" "Abril" ...
#>  $ dia_egr   : num [1:1000] 5 2 17 13 9 5 17 18 21 19 ...
#>   ..- attr(*, "label")= chr " Día egreso"
#>   ..- attr(*, "format.spss")= chr "F2.0"
#>  $ fecha_egr : chr [1:1000] "2020/09/05 00:00:00.000" "2020/11/02 00:00:00.000" "2020/02/17 00:00:00.000" "2020/08/13 00:00:00.000" ...
#>   ..- attr(*, "label")= chr " Fecha de egreso"
#>   ..- attr(*, "format.spss")= chr "A69"
#>   ..- attr(*, "display_width")= int 23
#>  $ dia_estad : num [1:1000] 6 3 3 1 3 1 1 1 6 1 ...
#>   ..- attr(*, "label")= chr " Días estada"
#>   ..- attr(*, "format.spss")= chr "F6.0"
#>  $ con_egrpa : dbl+lbl [1:1000] 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 3, 1, 1, 1, 1, 1,...
#>    ..@ label      : chr " Condición del egreso"
#>    ..@ format.spss: chr "F1.0"
#>    ..@ labels     : Named num [1:3] 1 2 3
#>    .. ..- attr(*, "names")= chr [1:3] "Vivo" "Fallecido menos de 48 horas" "Fallecido en 48 horas y más"
#>  $ esp_egrpa : dbl+lbl [1:1000] 37, 23, 41, 21,  7, 27, 47, 21, 29, 27, 30, 21, 27, 2...
#>    ..@ label      : chr " Especialidad del egreso"
#>    ..@ format.spss: chr "F2.0"
#>    ..@ labels     : Named num [1:52] 1 2 3 4 5 6 7 8 9 10 ...
#>    .. ..- attr(*, "names")= chr [1:52] "Alergología" "Atención Primaria de la Salud" "Cardiología" "Cardiopediatría" ...
#>  $ cau_cie10 : chr+lbl [1:1000] C169 , U072 , A090 , O809 , K469 , K802 , S903 , O730...
#>    ..@ label        : chr "Causa de lista internacional detallada a 4 dígitos de la CIE-10"
#>    ..@ format.spss  : chr "A15"
#>    ..@ display_width: int 5
#>    ..@ labels       : Named chr [1:18445] "A00" "B00" "C00" "D00" ...
#>    .. ..- attr(*, "names")= chr [1:18445] "Cólera" "Infecciones herpética [herpes simple]" "Tumor maligno del labio" "Carcinoma in situ de la cavidad bucal, del esófago y del estómago" ...
#>  $ cant_res  : chr+lbl [1:1000] 0901, 0901, 0901, 1207, 0901, 0901, 0901, 1701, 0901,...
#>    ..@ label        : chr "SecA Cantón de residencia habitual del paciente"
#>    ..@ format.spss  : chr "A12"
#>    ..@ display_width: int 14
#>    ..@ labels       : Named chr [1:222] "8800" "0110" "1110" "1210" ...
#>    .. ..- attr(*, "names")= chr [1:222] "Exterior" "Oña" "Puyango" "Buena Fé" ...
#>  $ parr_res  : chr+lbl [1:1000] 090150, 090104, 090150, 120750, 090150, 090112, 09015...
#>    ..@ label        : chr "SecA Parroquia de residencia habitual del paciente"
#>    ..@ format.spss  : chr "A18"
#>    ..@ display_width: int 20
#>    ..@ labels       : Named chr [1:1308] "880000" "010110" "170110" "090110" ...
#>    .. ..- attr(*, "names")= chr [1:1308] "Exterior" "San Blas" "El Condado" "Rocafuerte" ...
#>  $ causa3    : chr+lbl [1:1000] C16, U07, A09, O80, K46, K80, S90, O73, P36, J90, J44...
#>    ..@ label      : chr "Causa de lista internacional detallada a 3 dígitos de la CIE-10"
#>    ..@ format.spss: chr "A9"
#>    ..@ labels     : Named chr [1:2054] "A00" "B00" "C00" "D00" ...
#>    .. ..- attr(*, "names")= chr [1:2054] "Cólera" "Infecciones herpética [herpes simple]" "Tumor maligno del labio" "Carcinoma in situ de la cavidad bucal, del esófago y del estómago" ...
#>  $ cap221rx  : dbl+lbl [1:1000]  2, 22,  1, 15, 11, 11, 19, 15, 16, 10, 10, 15, 22, 2...
#>    ..@ label        : chr "Capitulo lista 221"
#>    ..@ format.spss  : chr "F4.0"
#>    ..@ display_width: int 10
#>    ..@ labels       : Named num [1:22] 1 2 3 4 5 6 7 8 9 10 ...
#>    .. ..- attr(*, "names")= chr [1:22] "I Ciertas enfermedades infecciosas y parasitarias (A00 -B99)" "II Neoplasias (C00-D48)" "III Enfermedades de la sangre y de los órganos hematopoyéticos y otros tras. que afectan el mecan. de la inmu (D50-D89)" "IV Enfermedades endocrinas, nutricionales y metabólicas  (E00-E90)" ...
#>  $ cau221rx  : dbl+lbl [1:1000]  23, 222,   1, 160, 114, 119, 205, 159, 167, 109, 105...
#>    ..@ label        : chr "Lista especial de 221 grupos"
#>    ..@ format.spss  : chr "F3.0"
#>    ..@ display_width: int 10
#>    ..@ labels       : Named num [1:222] 1 2 3 4 5 6 7 8 9 10 ...
#>    .. ..- attr(*, "names")= chr [1:222] "Enfermedades infecciosas intestinales (A00-A09)" "Tuberculosis (A15-A19)" "Ciertas zoonosis bacterianas (A20-A28)" "Otras enfermedades bacterianas (A30-A49)" ...
#>  $ cau298rx  : dbl+lbl [1:1000]  60, 299,   5, 243, 188, 195, 281, 242, 250, 179, 175...
#>    ..@ label        : chr "Lista de causas 298"
#>    ..@ format.spss  : chr "F3.0"
#>    ..@ display_width: int 10
#>    ..@ labels       : Named num [1:299] 1 2 3 4 5 6 7 8 9 10 ...
#>    .. ..- attr(*, "names")= chr [1:299] "Cólera" "Fiebres tifoidea y paratifoidea" "Shigelosis" "Amebiasis" ...