População dos municípios brasileiros, 2000-2021

Estimativas populacionais anuais por sexo e faixa etária quinquenal para os municípios brasileiros, de 2000 a 2021, calculadas antes do Censo 2022. Modifica a tabela de dados resultante da função mun_sex_pop do pacote brpop (https://github.com/rfsaldanha/brpop), permitindo seu uso como um banco de dados, além de mudar o rótulo de faixas etárias e categorias de sexo(v. Value). Exige a instalação prévia dos pacotes brpop e dplyr. Essas estimativas foram posteriormente atualizadas e são mantidas aqui apenas para fins de registro e comparação com as medidas e estimativas atuais. (V. Details)

Usage

popbr2000_2021(
  anoi = NULL,
  anof = NULL,
  uf = NULL,
  munic = NULL,
  droplevels = TRUE
)

Arguments

anoi

Ano de início do período a ser considerado, de 2000 a 2021. Argumento opcional, se não for preenchido, são consideradas as estimativas para todos os anos disponíveis, i.e., de 2000 a 2021 (v. Examples).

anof

Ano de fim do período a ser considerado, de 2000 a 2021. Argumento opcional, se não for preenchido, são consideradas as estimativas para todos os anos disponíveis, i.e., de 2000 a 2021 (v. Examples).

uf

Vetor com a(s) sigla(s) da(s) Unidade(s) da Federação de interesse.

munic

Vetor com o código IBGE do(s) municípios(s) de interesse.

droplevels

Vetor lógico. Se TRUE (padrão), exclui as categorias ("levels") sem observações em vetores da classe factor, comportamento normalmente desejável em seleções por município, UF ou região.

Value

Um banco de dados de classes data.table e data.frame com oito variáveis: CO_UF, UF_SIGLA, REGIAO, ano, mun, sexo, fxetar5, pop. A variável sexo é um factor com dois levels: "masc" e "fem". A variável fxetar5 representa a idade agrupada em 17 categorias – 16 faixas quinquenais (0-4, ... 75-79) e a última aberta a partir dos 80 anos (80 e +).

Details

Até a recente publicação dos dados do Censo 2022 e após um hiato na publicação de estimativas populacionais, em que o último arquivo de dados publicado no DATASUS era da população em 2012, trabalhávamos com as "Estimativas populacionais anuais por sexo e faixa etária quinquenal para os municípios brasileiros, de 2000 a 2021", que podiam ser tabuladas no TABNET mas cujos arquivos de dados não estavam disponíveis.

Em 2022 Raphael Saldanha dispôs-se ao trabalho de fazer as muitas tabulações necessárias e criou o pacote brpop, com as as estimativas da população por sexo e faixa etária para os municípios brasileiros, de 2000 a 2021. Entretanto, as tabelas em brpop tinham (têm) o total (a soma da população nas diferentes faixas etárias), e os rótulos das faixas etárias são longos e estavam (estão) em inglês, por isso criei outra função (popbr2000_2021) que retornasse a população com os rótulos em português e apenas com a população estimada em cada faixa etária (sem o total).

Com o Censo 2022 viu-se que os resultados estavam superestimados e as informações foram atualizadas e o período expandido até 2024. Essas novas informações voltaram a ser publicadas em arquivos no servidor FTP do DATASUS, facilitando seu uso como banco de dados, sem exigir a tabulação no TABNET.

Assim, a função popbr2000_2021 não é mais necessária e não deve ser usada em novos estudos, a não ser para análises metodológicas e comparação de resultados com as medidas atualizadas. Por esse motivo ela é mantida no pacote.

References

Nota técnica atualização: http://tabnet.datasus.gov.br/cgi/IBGE/SEI_MS-0034745983-Nota_Tecnica_final.pdf Nota técnica do estudo original: http://tabnet.datasus.gov.br/cgi/IBGE/NT-POPULACAO-RESIDENTE-2000-2021.PDF Essas informações podem ser tabuladas em http://tabnet.datasus.gov.br/cgi/deftohtm.exe?ibge/cnv/popsvsbr.def.

See also

Examples

library(dplyr)
# Ano 2021, todos os municípios brasileiros:
popbr2000_2021(2021)
#> Setting `max_tries = 2`.
#> Setting `max_tries = 2`.
#> Key: <CO_UF>
#>          CO_UF UF_SIGLA REGIAO   ano code_muni   sexo fxetar5   pop
#>         <char>   <fctr> <fctr> <num>    <char> <fctr>  <fctr> <int>
#>      1:     11       RO      N  2021    110001   masc     0-4   819
#>      2:     11       RO      N  2021    110001   masc     5-9   822
#>      3:     11       RO      N  2021    110001   masc   10-14   802
#>      4:     11       RO      N  2021    110001   masc   15-19   832
#>      5:     11       RO      N  2021    110001   masc   20-24   904
#>     ---                                                            
#> 189376:     53       DF     CO  2021    530010    fem   60-64 68202
#> 189377:     53       DF     CO  2021    530010    fem   65-69 52371
#> 189378:     53       DF     CO  2021    530010    fem   70-74 37881
#> 189379:     53       DF     CO  2021    530010    fem   75-79 24459
#> 189380:     53       DF     CO  2021    530010    fem  80 e + 28485
# Anos 2019 a 2021, RS:
popbr2000_2021(2019, 2021, uf = "RS") %>%
  dplyr::group_by(ano) %>%
  dplyr::summarise(pop = sum(pop))
#> # A tibble: 3 × 2
#>     ano      pop
#>   <dbl>    <int>
#> 1  2019 11377239
#> 2  2020 11422973
#> 3  2021 11466630
# Anos 2000 a 2003, AC:
popbr2000_2021(anof = 2003, uf = "AC") %>%
  group_by(ano) %>%
  summarise(pop = sum(pop))
#> # A tibble: 4 × 2
#>     ano    pop
#>   <dbl>  <int>
#> 1  2000 582172
#> 2  2001 600611
#> 3  2002 619261
#> 4  2003 637959
# Anos 2014 a 2016, Cerro Largo, RS:
popbr2000_2021(2014, 2016, munic = "430520") %>%
  group_by(sexo, fxetar5) %>%
  summarise(pop = sum(pop))
#> `summarise()` has grouped output by 'sexo'. You can override using the
#> `.groups` argument.
#> # A tibble: 34 × 3
#> # Groups:   sexo [2]
#>    sexo  fxetar5   pop
#>    <fct> <fct>   <int>
#>  1 masc  0-4      1216
#>  2 masc  5-9      1295
#>  3 masc  10-14    1338
#>  4 masc  15-19    1512
#>  5 masc  20-24    1597
#>  6 masc  25-29    1719
#>  7 masc  30-34    1616
#>  8 masc  35-39    1497
#>  9 masc  40-44    1421
#> 10 masc  45-49    1470
#> # ℹ 24 more rows