Baixa os arquivos da AIH e classifica as internações em CSAP

Descarrega os "arquivos da AIH" (arquivos RD<UFAAMM>.DBC das Bases de Dados do Sistema de Informações Hospitalares do SUS - BD-SIH/SUS) do site FTP do DATASUS e classifica as internações segundo a Lista Brasileira de Condições Sensíveis à Atenção Primária.

Usage

fetchcsap(
  uf = "all",
  anoinicio,
  mesinicio = 1,
  anofim = NULL,
  mesfim = NULL,
  periodo = "interna",
  cep = FALSE,
  cnes = FALSE,
  ...
)

Arguments

uf

Unidade da Federação. A sigla da UF ou um vetor com as siglas das UF de interesse, entre aspas e em letras maiúsculas. Para todo o Brasil (padrão), use "all".

anoinicio

Ano de competência da AIH para início da seleção dos dados, em formato numérico; sem padrão.

mesinicio

Mês de competência da AIH para início da seleção dos dados, em formato numérico; por padrão é 1.

anofim

Ano de competência da AIH para fim da seleção dos dados, em formato numérico; por padrão é igual ao ano seguinte ao ano de início (anoinicio + 1).

mesfim

Mês de competência da AIH para fim da seleção dos dados, em formato numérico; por padrão é 6 (junho). V. detalhes.

periodo

O período definido refere-se ao mês e ano de "competência" da AIH ou à data de internação? O padrão ("interna") é a internação. V. detahes.

cep

CEP de internação

cnes

CNES do estabelecimento que gerou a AIH

...

Permite o uso de outros parâmetros de csapAIH

Value

Um objeto de classes data.table e data.frame com as seguintes variáveis:

munres Município de residência do paciente

munint Município de internação do paciente

sexo Sexo do paciente

idade Idade do paciente em anos completos

fxetar5 Faixa etária quinquenal (0-4, ..., 76-79, 80 e +)

csap Internação por CSAP (sim/não)

grupo Grupo de causa da Lista Brasileira de ICSAP, ou "não-CSAP"

cid Diagnóstico principal da internação, segundo a Classificação Internacional de Doenças, 10ª Revisão

data.inter Data da internação

data.saida Data da alta

Details

- Período de download dos arquivos e de internação dos sujeitos.

Os "arquivos da AIH" são definidos por mês e ano de "competência", e não da data de internação. Assim, o arquivo de um determinado "mês de competência" pode incluir registros de internações ocorridas em outro mês ou ano, enquanto pode não incluir todos os casos ocorridos naquele mês. Por padrão, fetchcsap usa o argumento periodo = "interna" para selecionar os casos por data de internação de acordo com o período definido nos argumentos anoinicio, mesinicio e anofim, de modo a iniciar no primeiro dia do ano e mês de competência (anoinicio e mesinicio) e terminar em 31 de dezembro do ano anterior ao definido em anofim.

Assim, por padrão, a função exige apenas a definição do ano de início dos casos. Se o usuário definir apenas esse argumento, fetchcsap fará o download e leitura dos arquivos (de todo o Brasil) de todos os meses até junho do ano seguinte para então selecionar as internações ocorridas no ano definido em anoinicio.

- fetchcsap é apenas uma abreviatura para um uso específico da função fetch_datasus, do pacote microdatasus, de Raphael Saldanha. Funciona apenas com o SIH/SUS, através do argumento information_system = "SIH-RD", e faz apenas o download das variáveis exigidas pela função csapAIH, i.e., DIAG_PRINC, NASC, DT_INTER, DT_SAIDA, IDADE, COD_IDADE, MUNIC_RES, MUNIC_MOV, SEXO, N_AIH, PROC_REA, IDENT, CEP, CNES.

See also

Examples

# Internações de todo o Brasil, ocorridas no ano de 2023 e registradas até jun/2024:
# Colocando apenas o ano, único argumento obrigatório, a função executa o
# download dos arquivos RD??????.DBC de todas as UF de jan/2023 a jun/2024 e
# então extrais apenas os registros com data de internação em 2023.
if (FALSE) { # \dontrun{
  fetchcsap(anoinicio = 2023)
} # }
# Diferença entre o mês e ano de "competência" da AIH e a data de internação da pessoa,
# exemplo com as internações no Acre, registradas no mês de competência jan 2023:
ac.comp <- fetchcsap("AC", 2023, periodo = 'competencia')
#> ℹ Your local Internet connection seems to be ok.
#> ℹ DataSUS FTP server seems to be up and reachable.
#> ℹ Starting download...
#> Importados 4.165 registros.
#> Excluídos 1.179 (28,3%) registros de procedimentos obstétricos.
#> Excluídos 1 (0%) registros de AIH de longa permanência.
#> Exportados 2.985 (71,7%) registros.
nrow(ac.comp)
#> [1] 2985
summary(ac.comp$data.inter)
#>         Min.      1st Qu.       Median         Mean      3rd Qu.         Max. 
#> "2022-07-17" "2022-11-27" "2022-12-26" "2022-12-14" "2023-01-10" "2023-01-31" 
ac.int <- fetchcsap("AC", 2023, mesfim = 1, anofim = 2023)
#> ℹ Your local Internet connection seems to be ok.
#> ℹ DataSUS FTP server seems to be up and reachable.
#> ℹ Starting download...
#> Importados 1.917 registros.
#> Excluídos 672 (35,1%) registros de procedimentos obstétricos.
#> Excluídos NA (NA%) registros de AIH de longa permanência.
#> Exportados 1.245 (64,9%) registros.
nrow(ac.int)
#> [1] 1245
summary(ac.int$data.inter)
#>         Min.      1st Qu.       Median         Mean      3rd Qu.         Max. 
#> "2023-01-01" "2023-01-06" "2023-01-12" "2023-01-12" "2023-01-18" "2023-01-31" 
# Assim, há
nrow(ac.comp) - nrow(ac.int)
#> [1] 1740
# internações registradas naquele mês de competência, mas que ocorreram em outro momento.