Faixa etária em três grandes categorias

Recodifica vetores com a idade em números inteiros ou em faixas etárias quinquenais em três grandes faixas etárias: 0-14, 15-59 e 60 e + anos.

Usage

fxetar3g(idade = NULL)

Arguments

idade

Vetor representando a idade ou a faixa etária. Pode ser numérico, em valores contínuos ou inteiros, ou da classe factor ou character representando 17 faixas etárias quinquenais, rotuladas conforme o resultado da função fxetar_quinq (números separados por hífen, sem espaços: "0-4", ..., "75-79", "80 e +").

Value

Se fornecida a idade, devolve um fator com as frequências observadas em cada faixa etária; se fornecida a faixa etária, devolve um vetor da classe caractere com as frequências de cada faixa etária; se não é informado o argumento, a função devolve um vetor com os nomes das faixas etárias. V. exemplos.

Examples

# Apenas citar os grupos:
fxetar3g()
#> [1] "0-14"  "15-59" "60e+" 

# Categorizar a idade
## Criar um vetor para idade
idade <- as.integer(runif(100, 0, 100))
## Computar a faixa etária
fxetar3g(idade) |> table()
#> 
#>  0-14 15-59  60e+ 
#>    13    53    34 

# Recategorizar a faixa etária quinquenal e
# computar a faixa etária em três grandes gupos
fxetar_quinq(idade) |>
fxetar3g() |> table()
#> 
#>  0-14 15-59  60e+ 
#>    13    53    34